requirements.txt
# app.py: 1
# auth.py: 1
Flask == 1.0.2



# action_detection.py: 2
# preprocessing.py: 1
pandas == 0.23.4

# action_detection.py: 3
# preprocessing.py: 3
scikit_learn == 0.20.0
mysql-connector-python == 8.0.15

xgboost == 0.6a2

waitress == 1.1.0